console: Move portal component
